third_party/WebKit/LayoutTests/inspector/elements/styles-3/styles-disable-property-after-selector-edit-expected.txt

Verifies that sequence of setting selector and disabling property works.

Red text here.

Running: selectInitialNode

Running: editSelector

Running: testDisableProperty


#### AFTER PROPERTY DISABLED ####


[expanded] 
element.style { ()

[expanded] 
#inspected, .INSERTED-OTHER-SELECTOR { (styles-disable-…r-edit.html:36 -> styles-disable-property-after-selector-edit.html:36:39)
/-- overloaded --/ /-- disabled --/     /* color: red; */

[expanded] 
div { (user agent stylesheet)
    display: block;
